Add missing logos from gilbarbara repository fixing script to list all files

+++ svgsFolderUrl2plantUmlSprites.groovy
@@ -32,8 +32,9 @@ final PNGS_DIR = new File('pngs')
 PNGS_DIR.mkdirs()
 final SPRITES_LISTING = new SpritesListing(new File('sprites-list.md'), PNGS_DIR)
 
-def cli = new CliBuilder(usage: "${this.class.getSimpleName()}.groovy [options] <svgs URL>", stopAtNonOption: false, footer: "Usage example: ./${this.class.getSimpleName()}.groovy https://api.github.com/repos/gilbarbara/logos/contents/logos")
+def cli = new CliBuilder(usage: "${this.class.getSimpleName()}.groovy [options] <svgs URL>", stopAtNonOption: false, footer: "Usage example: ./${this.class.getSimpleName()}.groovy https://github.com/gilbarbara/logos/tree/master/logos")
 cli.s(longOpt: 'scale', args: 1, argName: 'scale', "Scale (eg: 0.5 to reduce size to half) of generated sprites. Default value: $DEFAULT_SCALE")
+cli.c(longOpt: 'use-cache', "When specified, already downloaded files are not re-downloaded")
 def options = cli.parse(args)
 !options && System.exit(1)
 if (!options.arguments()) {
@@ -48,30 +49,63 @@ if (options.arguments().size() > 1) {
 }
 def svgsUrl = options.arguments()[0]
 def scaleFactor = options.s ?: DEFAULT_SCALE
+def useCache = options.c
 
 listSvgsUrls(svgsUrl)
   .collect {
-    downloadFile(it, TMP_DIR)
+    downloadFile(it, TMP_DIR, useCache)
   }
   .collect {
     svg2Png(it, PNGS_DIR)
   }
   .collect {
-    png2PlantUmlSprite(it, scaleFactor, SPRITES_DIR)
+    scaleImage(it, scaleFactor)
+  }
+  .collect {
+    png2PlantUmlSprite(it, SPRITES_DIR)
   }
   .each {
     SPRITES_LISTING.addSprite(it)
   }
 
 def listSvgsUrls(baseUrl) {
-  new JsonSlurper()
-    .parseText(new URL(baseUrl).text)
-    .collect{ it['download_url'] }
+  def matcher = baseUrl =~ /^https:\/\/github.com\/([^\/]+)\/([^\/]+)\/tree\/(.*)$/
+  if (!matcher.matches()) {
+    throw new IllegalArgumentException("Provided URL is not a GitHub folder URL")
+  }
+  def owner = matcher[0][1]
+  def repo = matcher[0][2]
+  def path = matcher[0][3]
+
+  def treeUrl = findTreeUrl("https://api.github.com/repos/${owner}/${repo}/git/trees", path.split('/'))
+  return getUrlJson(treeUrl).tree.collect {
+    "https://raw.githubusercontent.com/${owner}/${repo}/${path}/${it.path}"
+  }
 }
 
-def downloadFile(url, workDir) {
+def findTreeUrl(treesBaseUrl, pathParts) {
+  def ret = "${treesBaseUrl}/${pathParts[0]}"
+  for (i=1; i<pathParts.size(); i++) {
+    def pathNode = getUrlJson(ret).tree.find { it.path == pathParts[i] }
+    if (!pathNode) {
+      throw new IllegalArgumentException("Not found provided path in GitHub")
+    }
+    ret = pathNode.url
+  }
+  return ret
+}
+
+def getUrlJson(url) {
+  return new JsonSlurper()
+      .parseText(new URL(url).text)
+}
+
+def downloadFile(url, workDir, useCache) {
   def fileName = Paths.get(new URI(url).path).fileName
   def svgFile = new File("$workDir/${fileName}")
+  if (useCache && svgFile.exists()) {
+    return svgFile
+  }
   svgFile.delete()
   svgFile << new URL(url).text
   return svgFile
@@ -88,10 +122,24 @@ def svg2Png(svg, workDir) {
   return pngFile
 }
 
-def png2PlantUmlSprite(png, scaleFactor, outputDir) {
+def scaleImage(png, scaleFactor) {
+  BufferedImage im = ImageIO.read(png)
+  def width = (im.width * scaleFactor) as Integer
+  def height = (im.height * scaleFactor) as Integer
+  BufferedImage scaledImage = new BufferedImage(width, height, im.type)
+  Graphics2D graphics2D = scaledImage.createGraphics()
+  graphics2D.setRenderingHint(RenderingHints.KEY_INTERPOLATION, RenderingHints.VALUE_INTERPOLATION_BILINEAR)
+  graphics2D.setRenderingHint(RenderingHints.KEY_RENDERING, RenderingHints.VALUE_RENDER_QUALITY)
+  graphics2D.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON)
+  graphics2D.drawImage(im, 0, 0, width, height, null)
+  graphics2D.dispose()
+  ImageIO.write(scaledImage, 'png', png);
+  return png
+}
+
+def png2PlantUmlSprite(png, outputDir) {
   BufferedImage im = ImageIO.read(png)
   removeAlpha(im)
-  im = scaleImage(im, scaleFactor)
   String spriteName = png.name.replace(".png","")
   def spriteFile = new File("$outputDir/${spriteName}.puml")
   spriteFile.delete()
@@ -108,19 +156,6 @@ def removeAlpha(im) {
   finally {
     graphics.dispose()
   }
-}
-
-def scaleImage(im, scaleFactor) {
-  def width = (im.width * scaleFactor) as Integer
-  def height = (im.height * scaleFactor) as Integer
-  BufferedImage ret = new BufferedImage(width, height, im.getType());
-  Graphics2D graphics2D = ret.createGraphics();
-  graphics2D.setRenderingHint(RenderingHints.KEY_INTERPOLATION, RenderingHints.VALUE_INTERPOLATION_BILINEAR);
-  graphics2D.setRenderingHint(RenderingHints.KEY_RENDERING, RenderingHints.VALUE_RENDER_QUALITY);
-  graphics2D.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
-  graphics2D.drawImage(im, 0, 0, width, height, null);
-  graphics2D.dispose();
-  return ret;
 }
 
 class SpritesListing {
